ER-219 :construction: - Moving Dashboard and VacancySummary to Projections namespace inside vacancies client.
