+++ src/Employer.Web/Orchestrators/DashboardOrchestrator.cs
@@ -1,6 +1,6 @@
 ﻿using Esfa.Recruit.Employer.Web.Services;
 using Esfa.Recruit.Employer.Web.ViewModels;
-using Esfa.Recruit.Vacancies.Client.Domain.Entities;
+using Esfa.Recruit.Vacancies.Client.Domain.Projections;
 using Esfa.Recruit.Vacancies.Client.Domain.QueryStore;
 using SFA.DAS.EAS.Account.Api.Types;
 using System.Collections.Generic;
