+++ src/Employer.Web/ViewModels/DashboardViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Esfa.Recruit.Vacancies.Client.Domain.Entities;
+using Esfa.Recruit.Vacancies.Client.Domain.Projections;
 using System.Collections.Generic;
 
 namespace Esfa.Recruit.Employer.Web.ViewModels
