+++ src/Employer.Web/Views/Dashboard/Index.cshtml
@@ -27,7 +27,7 @@
 			{
 				<tr>
 					<td><a asp-route="@RouteNames.Preview_Index_Get" asp-route-vacancyId="@vacancy.Id">@vacancy.Title</a></td>
-					<td></td>
+					<td>@vacancy.Status</td>
 				</tr>
 			}
 			</tbody>
