+++ src/Recruit.Vacancies.Client/Domain/Projections/Dashboard.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
+namespace Esfa.Recruit.Vacancies.Client.Domain.Projections
 {
     public class Dashboard
     {
