+++ src/Recruit.Vacancies.Client/Domain/Projections/VacancySummary.cs
@@ -1,7 +1,7 @@
 ﻿using Esfa.Recruit.Vacancies.Client.Domain.Enums;
 using System;
 
-namespace Esfa.Recruit.Vacancies.Client.Domain.Entities
+namespace Esfa.Recruit.Vacancies.Client.Domain.Projections
 {
     public class VacancySummary
     {
